Website Design Push
A big update to the Website's U.I. design and layout.

A header has been added and darkened so it seems more like a banner/header than just an image. Text description of the projects added to the initial project view. Outside of that the actual project pages have been changed to just have the role I had on the project, the video links, and the download links. The videos embedded into the site have had their resolutions changed so they seem less stretched out.

Still need to complete the Email Contact Forum however.